Refactor sendEndGameMessage to GameSession
