+++ src/test/ci/game_session.test.ts
@@ -127,7 +127,7 @@ describe("game session", () => {
 
                 // end session
                 const sendEndGameMessageStub = sandbox.stub(
-                    discord_utils,
+                    gameSession,
                     "sendEndGameMessage"
                 );
